requirements.txt
alembic==1.0.9
bcrypt==3.1.6
coverage==4.5.3
decorator==4.4.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==19.9.0
importlib-metadata==0.18
jsonschema==3.0.1
py==1.8.0
PyMySQL==0.9.3
pytest==4.6.3
pytest-cov==2.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
requests==2.21.0
sentry-sdk==0.7.12
simplejson==3.16.0
SQLAlchemy==1.3.3
ua-parser==0.8.0
user-agents==2.0
Werkzeug==0.15.2
